major refactoring of the layout and method names.
