+++ MIT-LICENSE
@@ -1,4 +1,5 @@
-Copyright (c) 2008 cracklabs.com
+Copyright on Updates - Copyright (c) 2011 OctopusApp Inc. (http://getjobber.com), released under the MIT license
+Original Copyright - Copyright (c) 2008 cracklabs.com, released under the MIT license
 
 Permission is hereby granted, free of charge, to any person obtaining
 a copy of this software and associated documentation files (the
