+++ lib/prawnto.rb
@@ -5,6 +5,7 @@ require 'prawnto/railtie' if defined?(Rails)
 
 module Prawnto
   VERSION='0.1.0'
+  
   autoload :ActionControllerMixin, 'prawnto/action_controller_mixin'
   autoload :ActionViewMixin, 'prawnto/action_view_mixin'
   module TemplateHandlers
@@ -12,26 +13,25 @@ module Prawnto
     autoload :Dsl, 'prawnto/template_handlers/dsl'
   end
 
-  module TemplateHandler
-    autoload :CompileSupport, 'prawnto/template_handler/compile_support'
-  end
-
   class << self
     
-    def run_includes
-      ActionController::Base.send :include, Prawnto::ActionControllerMixin
-      ActionView::Base.send :include, Prawnto::ActionViewMixin
-    end
-    
-    def register_handlers
+    # Register the MimeType and the two template handlers.
+    def on_init
       Mime::Type.register "application/pdf", :pdf unless defined?(Mime::PDF)
       ActionView::Template.register_template_handler 'prawn', Prawnto::TemplateHandlers::Base
       ActionView::Template.register_template_handler 'prawn_dsl', Prawnto::TemplateHandlers::Dsl
     end
     
+    # Includes the mixins for ActionController and ActionView.
+    def on_load
+      ActionController::Base.send :include, Prawnto::ActionControllerMixin
+      ActionView::Base.send :include, Prawnto::ActionViewMixin
+    end
+    
+    # Runs the registration and include methods. This is used for testing only as the Railtie.rb usually runs these individually.
     def init_both
-      run_includes
-      register_handlers
+      on_init
+      on_load
     end
     
   end
